created a build-on-release-workflow.yml that contains the "Attach APK to release" action. Did some minor refactoring to the build-on-merge-workflow.yml and added a step to ensure the android sdk is created
These workflows should not only build an APK and upload it to a github artifact everytime we merge with main, but pull the artifact when we create a release and attach the latest build to the release.

This will allow for automated builds, and can even be connected to google play in the future so that when we make a release, it automatically releases it on the google play store.